+++ src/apis/productApi.js
@@ -15,7 +15,7 @@ export const updateProductImgApi = async (img) => {
 export const getProductLikeApi = async (userId, id) => {
     let response = null;
     try {
-        response = await instance.get("/product", { userId, id });
+        response = await instance.get("/user/product");
     } catch (e) {
         console.error(e);
         response = e.response
@@ -26,7 +26,7 @@ export const getProductLikeApi = async (userId, id) => {
 export const productLikeApi = async (productId) => {
     let response = null;
     try {
-        response = await instance.post("/product/like", {productId});
+        response = await instance.post("/user/product/like", {productId});
     } catch (e) {
         console.error(e);
         response = e.response
@@ -36,17 +36,18 @@ export const productLikeApi = async (productId) => {
 export const deleteProductDisLikeApi = async (productId) => {
     let response = null;
     try {
-        response = await instance.delete("/product/dislike", { productId });
+        response = await instance.delete("/user/product/dislike", { productId });
     } catch (e) {
         console.error(e);
         response = e.response
     }
 };
 
+// 확인
 export const basketAddProductApi = async (productId) => {
     let response = null;
     try {
-        response = await instance.post("/user/img", { productId });
+        response = await instance.post("/user/cart", { productId });
         if (response.status === 200) {
             alert('장바구니에 상품이 추가되었습니다!');
         }
@@ -58,16 +59,16 @@ export const basketAddProductApi = async (productId) => {
     return response;
 }
 
-
+// 확인
 // 장바구니 상품 가져오기
 export const getBasketProductsApi = async () => {
-    const response = await axios.get("/api/basket"); // 적절한 API 경로로 수정
+    const response = await axios.get("/user/cart"); // 적절한 API 경로로 수정
     return response.data; // 데이터 반환
 };
 
 // 상품 삭제 API
 export const deleteProductApi = async (productId) => {
-    await axios.delete(`/api/basket/${productId}`); // 적절한 API 경로로 수정
+    await axios.delete(`/user/cart/${productId}`); // 적절한 API 경로로 수정
 };
 
 export const buyProductApi = async (productIds) => {
